Refactor matrix samplers
