+++ duqtools/schema/_description_helpers.py
@@ -1,0 +1,7 @@
+from textwrap import dedent
+
+
+def formatter(s):
+    """Dedent and remove newlines."""
+    s = dedent(s)
+    return s.replace('\n', ' ').strip()
